Save Lavaan Plot 2

Make and save plots with lavaanPlot::lavaanPlot

saveLavaanPlot2(
  fit,
  filePath,
  coefs = TRUE,
  stand = TRUE,
  sig = 0.05,
  stars = "regress",
  width = 1000
)

Arguments

fit

A fitted lavaan::lavaan SEM model object (e.g., created using cfa() or sem()).

filePath

A character string with the path to save the plot, including the file name and extension (e.g., "path/to/file.png").

coefs

Logical. Whether to display path coefficients. Default is TRUE.

stand

Logical. Whether to use standardized coefficients. Default is TRUE.

sig

Numeric. The significance level to display. Default is 0.05.

stars

Character. If "regress", significant coefficients are marked with stars. Default is "regress".

width

Numeric. Width of the plot in pixels. Default is 1000.

Value

A saved plot of the SEM model as a PNG file.

Details

Saves a Structural Equation Model (SEM) plot generated by lavaanPlot::lavaanPlot. The function handles intermediate steps, including exporting the plot to SVG and converting to PNG.

See also

Author

Jasmijn Bazen j.l.bazen@students.uu.nl

Examples

if (FALSE) { # \dontrun{
Save_LavaanPlot2(fit = SEMModel1,
                filePath = "path/to/your/SEMModel.png",
                coefs = TRUE,
                stand = TRUE,
                sig = 0.05,
                stars = "regress",
                width = 1000)
} # }